characters and boxes on the screen all togethe rin random positions that do not coincide
